Reverting the pre-authentication process to correct logic:  only rely on authToken, nothing else, no fallbacks.
More work needed to test and fix the navigation issue that was originally addressed for  GCP-10518,  which caused this security issue.
Note that this change is just a starting point. Other fixes will be needed as we test the impact of this on overall security/pre-auth process. I expect more changes/fixes in other areas will still be needed.

Issue: BMS-1986
Reviewer: MatthewB