Use germplasm locationId to advance
Old code from manage Nursery had a harvest location combobox to select,
and some of the legacy code was there to set a hidden input name
harvestLocationId that will be sent as part of the request to advance
nursery/trial. This input was set in many different places based on
where the location was placed (when location could be added as condition
or environment). This frontend code was removed as there's no need to
keep all these places in sync now. Additionally, this was causing that
the location from the last environment was being selected to advance.

With this change, some of the backend code to receive a single
harvestLocation to set for all germplasm was kept in case we want to
bring it back in the future.
We copy the locationId calculated from measurement rows data in
dataProcessor.processPlotLevelData(advancingSourceCandidate, row)
in AdvancingSourceListFactory#createAdvancingSourceList to the
importedGermplasm object and then transfer it to the germplasm when
saving the list.
The code from
ChangeLocationExpressionDataProcessor.processEnvironmentLevelData is
kept though it's not exactly clear when it applies.

Issue: IBP-2161 / BMS-4612
Reviewer: None